Extract plot tabs to separate module
